+++ src/components/Inputs/TextField/stories/TextField.stories.tsx
@@ -9,15 +9,9 @@ const Template: ComponentStory<typeof TextField> = (args) => (
   </Stack>
 );
 
-const TemplatePicture: ComponentStory<typeof TextField> = (args) => (
+const TemplateFile: ComponentStory<typeof TextField> = (args) => (
   <Stack direction="row" spacing={2} alignItems="center" justifyContent="center" height="100%">
-    <TextField id="picture" label="Outlined" className="picture" {...args} />
-  </Stack>
-);
-
-const TemplatePictureVerticale: ComponentStory<typeof TextField> = (args) => (
-  <Stack direction="row" spacing={2} alignItems="center" justifyContent="center" height="100%">
-    <TextField id="picture-vertical" label="Outlined" className="picture-vertical" {...args} />
+    <TextField id="picture" {...args} />
   </Stack>
 );
 
@@ -36,14 +30,21 @@ Standard.args = {
   variant: "standard",
 };
 
-export const TypeFilePicture = TemplatePicture.bind({});
-TypeFilePicture.args = {
+export const File = Template.bind({});
+File.args = {
+  label: "",
+  type: "file",
+};
+
+export const FileWithLabel = TemplateFile.bind({});
+FileWithLabel.args = {
   label: "Ajouter une photo ou un fichier",
   type: "file",
 };
 
-export const TypeFilePictureVertical = TemplatePictureVerticale.bind({});
-TypeFilePictureVertical.args = {
+export const FileWithLabelDirectionColumn = TemplateFile.bind({});
+FileWithLabelDirectionColumn.args = {
+  dir: "column",
   label: "Ajouter une photo ou un fichier",
   type: "file",
 };